Setup absolute import
