+++ src/components/accordion/AccordionItem.tsx
@@ -1,11 +1,8 @@
 import { FC } from 'react';
-import { AccordionItemProps } from '../../types';
-import {
-  accordionClassesOptim,
-  AccordionItemClasses,
-} from '../../utils/accordion';
-import { AccordionContext } from '../../contexts/accordion';
-import { useAccordion } from '../../hooks/accordion';
+import { AccordionItemProps } from 'types/accordion';
+import { accordionClassesOptim, AccordionItemClasses } from 'utils/accordion';
+import { AccordionContext } from 'contexts/accordion';
+import { useAccordion } from 'hooks/accordion';
 
 const AccordionItem: FC<AccordionItemProps> = ({
   children,
@@ -25,4 +22,4 @@ const AccordionItem: FC<AccordionItemProps> = ({
   );
 };
 
-export default AccordionItem;
+export { AccordionItem };
