--- src/components/accordion/index.tsx
@@ -1,10 +1,0 @@
-import Accordion from './Accordion';
-import AccordionItem from './AccordionItem';
-import AccordionHeader from './AccordionHeader';
-import AccordionBody from './AccordionBody';
-
-export default Object.assign(Accordion, {
-  Item: AccordionItem,
-  Header: AccordionHeader,
-  Body: AccordionBody,
-});
