+++ src/contexts/accordion/index.ts
@@ -1,3 +1,3 @@
 import { createContext } from 'react';
-import { AccordionStatus } from '../../types';
+import { AccordionStatus } from 'types/accordion';
 export const AccordionContext = createContext<AccordionStatus>({});
