+++ src/pages/accordion/index.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import FAQ from 'examples/accordion/FAQ';
+const Accordions = () => {
+  return (
+    <>
+      <FAQ />
+    </>
+  );
+};
+
+export default Accordions;
